refactor: use glam types in projectile and presence_position
presence_position: remove redundant i32 intermediate in chunk_hash_base
(i32 as i64 as u64 → i32 as u64; sign-extension is identical).

projectile: replace manual component-wise ops with glam throughout:
- voxel_hit: per-component i64 coords → IVec3 via .as_ivec3(), 6-way
  bounds check → cmplt/cmpge, triple as u32 → .as_uvec3()
- voxel_hit: pivot/dims array construction via .map() + From<[T;3]>
- carve loop: scalar dx/dy/dz → IVec3 d, dot product sphere test,
  tuple (cx,cy,cz) → IVec3 c, component-wise bounds check
- hit_local/pivot_vec: reuse pivot_vec already in scope instead of
  repeating pivot[N] as f64; use hv.as_dvec3() + splat(0.5) - pivot_vec

+++ src/systems/presence_position.rs
@@ -86,9 +86,9 @@ fn splitmix64(mut h: u64) -> u64 {
 fn chunk_hash_base(world_seed: u64, chunk: IVec3) -> u64 {
     splitmix64(
         world_seed
-            .wrapping_add((chunk.x as i64 as u64).wrapping_mul(0x9e3779b97f4a7c15))
-            .wrapping_add((chunk.y as i64 as u64).wrapping_mul(0x6c62272e07bb0142))
-            .wrapping_add((chunk.z as i64 as u64).wrapping_mul(0x4d2c6dfc5ac42aad)),
+            .wrapping_add((chunk.x as u64).wrapping_mul(0x9e3779b97f4a7c15))
+            .wrapping_add((chunk.y as u64).wrapping_mul(0x6c62272e07bb0142))
+            .wrapping_add((chunk.z as u64).wrapping_mul(0x4d2c6dfc5ac42aad)),
     )
 }
